appli_image/requirements.txt
tensorflow==2.7.0
streamlit-drawable-canvas==0.8.0
streamlit==1.1.0
numpy==1.20.1
pillow==6.2.0
pandas==1.3.4
matplotlib==3.3.4